Remove exceptions : alt-x notably
alt are immediate keys. x is cut. so alt x is cut.
ctrl x is the delete prompt.

this makes necessary for user to know
how to invoke M-x, so later on we might
say binding M-x to something else than vanilla,
should imply some setup like
nu-rebind-m-x or so.