nu-set-control-mode
have a quick func to switch to pure cua
control+c copy, control+x cut, control+f find and so on.
Menus are now available using the Alt keys.

This is not recommended, but might be useful
whenever nu-mode is used altogether with other programs,
which are not keyboard friendly.